Simple Test
This test fails. More specifically, this test formats data in a way I don't really understand. It seems to be packing data out of order - which maybe has to do w/ bytes being variable length, but I doubt it. 

It also may be adding additional trailing bytes, which, I think, makes it a non-valid tx.